+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+## 0.24.0
+
+* `FEAT`: support custom category ([#168](https://github.com/camunda/element-templates-json-schema/pull/168))
+* `FEAT`: support `zeebe:script` binding ([#166](https://github.com/camunda/element-templates-json-schema/pull/166))
 * `FEAT`: support `zeebe:calledDecision` binding ([#165](https://github.com/camunda/element-templates-json-schema/pull/165))
 * `FEAT`: support `zeebe:formDefinition` binding ([#169](https://github.com/camunda/element-templates-json-schema/pull/169))